requirements.txt
line-bot-sdk==3.14.0
fastapi
uvicorn[standard]
google-genai>=1.49.0
aiohttp
pydantic
